更新

+++ 1_经验池/2_寄巧/无临时变量交换数.c
@@ -1,9 +1,9 @@
 #include<stdio.h>
 int main()
 {
-	int a,b;
+	float a,b;
 	printf("请输入两个树：");
-	scanf("%d%d",&a,&b);
+	scanf("%f%f",&a,&b);
 	a+=b;b=a-b;a-=b; // 交换a和b 而不使用临时变量
-	printf("\n%d,%d",a,b); 
+	printf("\n%f,%f",a,b); 
 }
